+++ src/App.css
@@ -1,7 +1,7 @@
-body {
-  margin-top: 5vh;
-  /* background-image: linear-gradient(135deg, #06beb6,#48b1bf); */
-  background-color: #bbbbbb;
+#home-btn {
+  position: absolute;
+  top: 2rem;
+  left: 2rem;
 }
 
 #writingArea, #previewArea {
