+++ src/components/common/homeButton/styles.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export const HomeButtonWrapper = styled.div`
+    position: absolute;
+    top: 2rem;
+    left: 2rem;
+`
+export const Button = styled.button`
+`
