+++ src/components/common/videoBackground/styles.js
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+export const VideoBackgroundWrapper = styled.div`
+    position: absolute;
+    overflow-y: hidden;
+    width: 100%;
+    height: 115%;
+    z-index: -1;
+`
+export const Video = styled.video`
+    width: 100%;
+    object-fit: cover;
+`
