+++ src/components/homework/styles.js
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+export const HomeworkWrapper = styled.div`
+    padding-top: 2rem;
+    padding-bottom: 2rem;
+`
